Clean up; Convert Flex to Grid
This commit removes unnecessary lines of code and converts all of the boxes from a flexbox layout into a grid. It also fixes all of the responsiveness issues and makes the site look slightly neater, especially for the smaller devices. By making the formatting the boxes in a grid, it makes it much easier to position the elements in a more organized and neat manner.